+++ cpp/src/db/insert/MemManagerImpl.h
@@ -18,10 +18,10 @@
 
 #pragma once
 
-#include "db/meta/Meta.h"
 #include "MemTable.h"
-#include "db/Status.h"
 #include "MemManager.h"
+#include "db/meta/Meta.h"
+#include "utils/Status.h"
 
 #include <map>
 #include <string>