+++ cpp/src/db/meta/Meta.h
@@ -15,12 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
+
 #pragma once
 
 #include "MetaTypes.h"
 #include "db/Options.h"
-#include "db/Status.h"
 #include "db/Types.h"
+#include "utils/Status.h"
 
 #include <cstddef>
 #include <memory>